+++ samples/Catalog/index.js
@@ -1,14 +1,9 @@
-//  Copyright © 2018-2021 PSPDFKit GmbH. All rights reserved.
-//
-//  THIS SOURCE CODE AND ANY ACCOMPANYING DOCUMENTATION ARE PROTECTED BY INTERNATIONAL COPYRIGHT LAW
-//  AND MAY NOT BE RESOLD OR REDISTRIBUTED. USAGE IS BOUND TO THE PSPDFKIT LICENSE AGREEMENT.
-//  UNAUTHORIZED REPRODUCTION OR DISTRIBUTION IS SUBJECT TO CIVIL AND CRIMINAL PENALTIES.
-//  This notice may not be removed from this file.
-//
+/**
+ * @format
+ */
 
-import { AppRegistry } from "react-native";
-import Catalog from "./Catalog";
-import Testing from './testing/Testing';
+import {AppRegistry} from 'react-native';
+import Catalog from './Catalog';
+import {name as appName} from './app.json';
 
-AppRegistry.registerComponent("Catalog", () => Catalog);
-AppRegistry.registerComponent('Testing', () => Testing);
+AppRegistry.registerComponent(appName, () => Catalog);