+++ samples/NativeCatalog/__tests__/App-test.js
@@ -1,0 +1,13 @@
+/**
+ * @format
+ */
+
+import 'react-native';
+import React from 'react';
+
+// Note: test renderer must be required after react-native.
+import renderer from 'react-test-renderer';
+
+it('renders correctly', () => {
+  // Add test here
+});